Rename a dir and shrink the test matrix
