+++ .github/workflows/ci.yml
@@ -11,8 +11,8 @@ jobs:
 
     strategy:
       matrix:
-        otp: [21.x, 22.x, 23.x]
-        elixir: [1.7.x, 1.8.x, 1.9.x, 1.10.x, 1.11.x]
+        otp: [21.x]
+        elixir: [1.8.x, 1.11.x]
 
     steps:
     - uses: actions/checkout@v2
@@ -42,8 +42,8 @@ jobs:
 
     strategy:
       matrix:
-        otp: [21.x, 22.x, 23.x]
-        elixir: [1.8.x, 1.9.x, 1.10.x, 1.11.x]
+        otp: [21.x]
+        elixir: [1.8.x, 1.11.x]
 
     services:
       db:
