+++ .tool-versions
@@ -1,2 +1,2 @@
-elixir 1.10.2
+elixir 1.7.4
 erlang 23.2.6
